Add support for the Indexer, part 1
Add support for the indexer from FRM2, which enumarates all devices
which are exposed from a TwinCAT system.
So far we support device types 5008 and 500c for motors.

The unit (like mm), and the name of the device are read out and pushed
into the EGU and DESC fields of the motorRecord.

The velocity, acceleration and soft limits are read as well.

Note that the indexer model has an own definition of a device state,
like IDLE, BUSY, ERROR.

In this commit the state machine is used to do absolute positioning.

Relative movements, move with constant velocity or homing are not
yet implemented, more work to be done.