Add support for the Indexer, part 1
Use a different device abstraction model (used at FRM2).
So far we can use a param device, which is an abstraction for
e.g. a motor.
Read out all devices from the Indexer, handle the motors only.
The poller is working, and point-to-point movements are supported.
Homing and "move with constant speed" are not yet supported.