Support the Indexer/PILS, part 1
Support the ADS based communication with TwinCAT based on this specification:
https://forge.frm2.tum.de/public/doc/plc/master/html/index.html

In short:
The IOC can read all the "devices" that are exposed by the TwinCAT PLC.
All devices with the device type 5008 (or 500c) are treated as motors,
the rest is ignored for the moment.
Parameters like acceleration may be exposed by the PLC per axis, or not.

The simulator exposes one axis and has been updated to handle the ADS
protocol.

Note thet this a kind of beta-release, tested under Debian 6 ppc
and Mac OS X.

More work is to be done:
- Support for 64 bit floating point values
- Get all values for all axes in one transaction in poll()
- Review the "connection closed" and re-etablished scenarios.
- Some cleanups here and there

Possible future work:
- Possibly implement the "symbol look up" via ADS
- Possibly make use of the notifications via ADS