netToDouble(), doubleToNet()m netToUint(), uintToNet()
Add converion functions to convert double and integers to and from the
an ADS message on the network.

The floating point values use always double inside EPICS and may
use float on the network.
The integer values use always "unsigned int" inside EPICS, and
may use 2, 4 or 8 byte on the network.
When using 8 bytes, only the lower 4 bytes are considered.
Parameters can use values > 2^31, which shouldn't be a problem in praxis.