Introduce a later poll - WIP
In order to update records and field in the motorRecord
(SPDB) try to start the poller later, after all records have been loaded.

This does not work as expected, as the ioc thread needs to do some work,
before the poller should poll.
In other words: A better synchronization is needed.