TC 200: fix timing problems
The test in 200_Ethercat-powerOff-JOG-CNEN-Error.py was often leaving
the real hardware in a non-usable state.
Power was off and the error was still there.

Add polling loops when the power is switched on or off
and add a loop when the -ErrRst bit is set.

While there, rename the test cases into 200x for consistancy.