Indexer: Support User limits
Major change to support the user limits for PILS,
add parameters in the IndexerAxis, handle it in the simulator
and, since we can not switch the limits off,
change motor_lib.py to set a very high/low value (which the driver
will read as "disable").
This makes TC122 pass.