Restore the noHomProc means no homed problem from v2.0.1
For "logical" axes in the TwinCAT system the "homed" status may be
"not homed".
EthercatMC v2.0.1 did treat this as a problem only when the
homing procedure was != 0.
That assumed that an axis with absolute encoders was always homed.

The later versions of the module does not have this assumption,
and simply looks at the "homed" flag for the axis.

This has lead to a regression, when an "old" TwinCAT SW should be
run against the latest EthercatMC.
Now the virtual axes like "Gap" and "Center" are no longer homed
and are put into alarm state.

There are 2 or 3 ways forward:
a) Restore the old handling
b) Fix the TwinCAT SW to report those axes as "homed"
c) Re-define EthercatMCHomProc_ as follows:
   When EthercatMCHomProc_ == -1, set EthercatMCHomProc_RB_ = 0
   and don't treat a not-homed axis as having an error.

My prefered solution would be (b).
In order to remove the regression, this implements (a)